rename accelerator class
